scripts/requirements.txt
google-api-python-client==2.83.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
requests==2.31.0
